ファームウェアロード・API互換性
P/Invoke呼び出し規約をStdCallに統一。テクニックのパラメータ順序を公式例に厳密化。Charge/DischargeのMethodParametersを実装。

+++ SequenceItems/DischargeSequenceItem.cs
@@ -176,23 +176,23 @@ public class DischargeSequenceItem : SequenceItem
       if (value is bool boolValue)
       {
         param.ParamType = (int)PARAM_TYPE.PARAM_BOOLEAN;
-        param.ParamVal = boolValue ? 1u : 0u;
+        param.ParamVal = boolValue ? 1 : 0;
       }
       else if (value is int intValue)
       {
         param.ParamType = (int)PARAM_TYPE.PARAM_INT;
-        param.ParamVal = (uint)intValue;
+        param.ParamVal = intValue;
       }
       else if (value is float floatValue)
       {
         param.ParamType = (int)PARAM_TYPE.PARAM_SINGLE;
-        param.ParamVal = BitConverter.ToUInt32(BitConverter.GetBytes(floatValue), 0);
+        param.ParamVal = BitConverter.ToInt32(BitConverter.GetBytes(floatValue), 0);
       }
       else if (value is double doubleValue)
       {
         param.ParamType = (int)PARAM_TYPE.PARAM_SINGLE;
         float f = (float)doubleValue;
-        param.ParamVal = BitConverter.ToUInt32(BitConverter.GetBytes(f), 0);
+        param.ParamVal = BitConverter.ToInt32(BitConverter.GetBytes(f), 0);
       }
 
       paramList.Add(param);
